Perception: fixed the missing namespace and include guards of kalman_filter_1d and moved it to common/math folder.
